pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "env-echo"
version = "1.0.0"
description = "Offline environment variable manager: generate, validate, mock, and audit .env files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "env-echo contributors"}]
keywords = ["env", "environment", "dotenv", "validation", "security", "audit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
env-echo = "env_echo.cli:main"

[tool.setuptools.packages.find]
include = ["env_echo*"]
